refactor: split supervisor diagnostics and recovery tests
