+++ .codex-supervisor/issue-journal.md
@@ -1,49 +1,37 @@
-# Issue #465: Test refactor: split pull-request-state tests by provider waits and state-policy coverage
+# Issue #463: Test refactor: split supervisor diagnostics and recovery coverage out of supervisor.test.ts
 
 ## Supervisor Snapshot
-- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/465
-- Branch: codex/issue-465
-- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-465
-- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-465/.codex-supervisor/issue-journal.md
-- Current phase: addressing_review
-- Attempt count: 4 (implementation=2, repair=2)
-- Last head SHA: 32df77eb6e9b894a2c5fc5871f75fe0e403edd54
+- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/463
+- Branch: codex/issue-463
+- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-463
+- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-463/.codex-supervisor/issue-journal.md
+- Current phase: local_review
+- Attempt count: 1 (implementation=1, repair=0)
+- Last head SHA: 3d5bf9a0dc52b8c4bb5e8832962ee96d3fa2fd50
 - Blocked reason: none
 - Last failure signature: none
 - Repeated failure signature count: 0
-- Updated at: 2026-03-17T09:01:08Z
+- Updated at: 2026-03-17T09:35:15Z
 
 ## Latest Codex Summary
-Addressed CodeRabbit thread `PRRT_kwDORgvdZ850zHi-` by removing the redundant `SupervisorConfig` type assertion in `src/pull-request-state-provider-waits.test.ts` and keeping the test assignment direct. Focused verification passed for the provider-waits suite and the project build.
-
-Summary: Fixed the redundant type assertion flagged in PR #471 and reverified the targeted suite plus build.
-State hint: addressing_review
-Blocked reason: none
-Tests: `npx tsx --test src/pull-request-state-provider-waits.test.ts`; `npm run build`
-Failure signature: none
-Next action: Commit and push the review fix to PR #471, then resolve the addressed review thread if no new feedback appears.
+- None yet.
 
 ## Active Failure Context
-- Category: review
-- Summary: No active local failure after applying the current review fix; waiting to push the update to PR #471.
-- Reference: https://github.com/TommyKammy/codex-supervisor/pull/471#discussion_r2945283266
-- Details:
-  - Resolved locally by replacing the casted assignment with `config.configuredBotSettledWaitSeconds = 3;` in `src/pull-request-state-provider-waits.test.ts` and removing the now-unused `SupervisorConfig` import.
+- None recorded.
 
 ## Codex Working Notes
 ### Current Handoff
-- Hypothesis: `pull-request-state` coverage is easier to maintain when provider wait behavior, merge/local-review policy, and same-head review-thread reprocessing are isolated into separate suites with a tiny shared fixture module.
-- What changed: Applied the PR #471 review fix by removing the redundant `SupervisorConfig` assertion in the provider-waits suite; the current local diff is limited to that cleanup plus this journal update.
+- Hypothesis: `supervisor.test.ts` shrinks safely if diagnostics (`doctor`/`status`/`explain`) and explicit reconciliation-recovery cases move into dedicated suites backed by a small supervisor-only fixture helper.
+- What changed: Added `src/supervisor/supervisor-test-helpers.ts`, extracted diagnostics coverage to `src/supervisor/supervisor-diagnostics.test.ts`, extracted reconciliation/recovery coverage to `src/supervisor/supervisor-recovery.test.ts`, and removed those cases from `src/supervisor/supervisor.test.ts`.
 - Current blocker: none
-- Next exact step: Commit and push the review-fix checkpoint on `codex/issue-465`, then clear the addressed review thread on PR #471 if GitHub shows the new head.
-- Verification gap: none for the review fix; the targeted suite and `npm run build` both passed locally.
-- Files touched: src/pull-request-state-test-helpers.ts; src/pull-request-state-provider-waits.test.ts; src/pull-request-state-policy.test.ts; src/pull-request-state-thread-reprocessing.test.ts; src/pull-request-state.test.ts; .codex-supervisor/issue-journal.md
-- Rollback concern: The helper is intentionally limited to deterministic fixture builders; pushing more policy logic into shared helpers would blur the suite boundaries this refactor is restoring.
+- Next exact step: Review the final diff, then commit the issue-463 refactor checkpoint on `codex/issue-463`.
+- Verification gap: none; the focused diagnostics/recovery suites, the remaining `supervisor.test.ts`, and `npm run build` all pass locally.
+- Files touched: src/supervisor/supervisor.test.ts; src/supervisor/supervisor-diagnostics.test.ts; src/supervisor/supervisor-recovery.test.ts; src/supervisor/supervisor-test-helpers.ts; .codex-supervisor/issue-journal.md
+- Rollback concern: Keep `supervisor-test-helpers.ts` limited to shared fixture setup for supervisor integration tests; if unrelated suites start depending on it, the ownership boundary this split created will blur again.
 - Last focused command: `npm run build`
 ### Scratchpad
 - Keep this section short. The supervisor may compact older notes automatically.
-- Reproducing signature before the fix: `src/pull-request-state.test.ts` was a single 1327-line, 38-test file mixing provider waits, settled waits, merge gating, local-review policy, and same-head review-thread reprocessing.
-- Verification commands: `npx tsx --test src/pull-request-state.test.ts`; `npx tsx --test src/pull-request-state-provider-waits.test.ts src/pull-request-state-policy.test.ts src/pull-request-state-thread-reprocessing.test.ts`; `npx tsx --test src/pull-request-state-*.test.ts`; `npm ci`; `npm run build`.
-- Local failure resolved: `npm run build` initially failed with `sh: 1: tsc: not found` because `node_modules` was absent in this worktree.
-- Draft PR: https://github.com/TommyKammy/codex-supervisor/pull/471
-- Review fix verification on 2026-03-17: `npx tsx --test src/pull-request-state-provider-waits.test.ts`; `npm run build`.
+- Reproducing signature before the fix: `src/supervisor/supervisor.test.ts` was a single 4851-line suite mixing diagnostics, reconciliation/recovery, and heavier execution-path coverage.
+- File size after extraction: `src/supervisor/supervisor.test.ts` is now 2670 lines; extracted suites are 1113 lines (`supervisor-diagnostics.test.ts`) and 813 lines (`supervisor-recovery.test.ts`).
+- Verification commands: `npx tsx --test src/supervisor/supervisor-diagnostics.test.ts`; `npx tsx --test src/supervisor/supervisor-recovery.test.ts`; `npx tsx --test src/supervisor/supervisor.test.ts`; `npm ci`; `npm run build`.
+- Local failure resolved: `npm run build` initially failed with `sh: 1: tsc: not found` because `node_modules` was absent in this worktree; `npm ci` restored the local toolchain.
