+++ src/supervisor/supervisor-recovery.test.ts
@@ -1,0 +1,813 @@
+import assert from "node:assert/strict";
+import fs from "node:fs/promises";
+import os from "node:os";
+import path from "node:path";
+import test from "node:test";
+import { GitHubIssue, GitHubPullRequest, IssueRunRecord, SupervisorStateFile } from "../core/types";
+import {
+  formatRecoveryLog,
+  reconcileMergedIssueClosures,
+  reconcileParentEpicClosures,
+  reconcileRecoverableBlockedIssueStates,
+  reconcileStaleActiveIssueReservation,
+  reconcileTrackedMergedButOpenIssues,
+} from "../recovery-reconciliation";
+import { shouldAutoRetryHandoffMissing } from "./supervisor-execution-policy";
+import { recoverUnexpectedCodexTurnFailure } from "./supervisor-failure-helpers";
+import { Supervisor } from "./supervisor";
+import {
+  branchName,
+  createConfig,
+  createRecord,
+  createSupervisorFixture,
+  executionReadyBody,
+  git,
+} from "./supervisor-test-helpers";
+
+test("reconcileRecoverableBlockedIssueStates requeues open handoff-missing issues without dropping repeat tracking", async () => {
+  const config = createConfig();
+  const original = createRecord();
+  const state: SupervisorStateFile = {
+    activeIssueNumber: null,
+    issues: {
+      "366": original,
+    },
+  };
+  const issues: GitHubIssue[] = [
+    {
+      number: 366,
+      title: "P3: Add regression coverage",
+      body: "",
+      createdAt: "2026-03-10T23:25:21Z",
+      updatedAt: "2026-03-10T23:25:21Z",
+      url: "https://example.test/issues/366",
+      state: "OPEN",
+    },
+  ];
+
+  let saveCalls = 0;
+  const stateStore = {
+    touch(record: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      return {
+        ...record,
+        ...patch,
+        updated_at: "2026-03-11T06:33:08.821Z",
+      };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  await reconcileRecoverableBlockedIssueStates(stateStore, state, config, issues, {
+    shouldAutoRetryHandoffMissing,
+  });
+
+  const updated = state.issues["366"];
+  assert.equal(updated.state, "queued");
+  assert.equal(updated.blocked_reason, null);
+  assert.equal(updated.last_error, null);
+  assert.equal(updated.codex_session_id, null);
+  assert.equal(updated.last_failure_signature, "handoff-missing");
+  assert.equal(
+    updated.last_failure_context?.summary ?? null,
+    "Codex completed without updating the issue journal for issue #366.",
+  );
+  assert.equal(updated.repeated_failure_signature_count, 1);
+  assert.equal(saveCalls, 1);
+});
+
+test("reconcileRecoverableBlockedIssueStates leaves closed issues blocked", async () => {
+  const config = createConfig();
+  const original = createRecord();
+  const state: SupervisorStateFile = {
+    activeIssueNumber: null,
+    issues: {
+      "366": original,
+    },
+  };
+  const issues: GitHubIssue[] = [
+    {
+      number: 366,
+      title: "P3: Add regression coverage",
+      body: "",
+      createdAt: "2026-03-10T23:25:21Z",
+      updatedAt: "2026-03-10T23:25:21Z",
+      url: "https://example.test/issues/366",
+      state: "CLOSED",
+    },
+  ];
+
+  let saveCalls = 0;
+  const stateStore = {
+    touch(record: IssueRunRecord): IssueRunRecord {
+      return record;
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  await reconcileRecoverableBlockedIssueStates(stateStore, state, config, issues, {
+    shouldAutoRetryHandoffMissing,
+  });
+
+  assert.deepEqual(state.issues["366"], original);
+  assert.equal(saveCalls, 0);
+});
+
+test("reconcileRecoverableBlockedIssueStates requeues requirements-blocked issues once metadata is execution-ready", async () => {
+  const config = createConfig();
+  const original = createRecord({
+    state: "blocked",
+    blocked_reason: "requirements",
+    last_error: "Missing required execution-ready metadata: scope, acceptance criteria, verification.",
+    last_failure_kind: null,
+    last_failure_context: {
+      category: "blocked",
+      summary: "Issue #366 is not execution-ready because it is missing: scope, acceptance criteria, verification.",
+      signature: "requirements:scope|acceptance criteria|verification",
+      command: null,
+      details: [
+        "missing_required=scope, acceptance criteria, verification",
+        "missing_recommended=depends on, execution order",
+      ],
+      url: "https://example.test/issues/366",
+      updated_at: "2026-03-11T01:50:41.997Z",
+    },
+    last_failure_signature: "requirements:scope|acceptance criteria|verification",
+    repeated_failure_signature_count: 2,
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: null,
+    issues: {
+      "366": original,
+    },
+  };
+  const issues: GitHubIssue[] = [
+    {
+      number: 366,
+      title: "P3: Add regression coverage",
+      body: executionReadyBody("Add regression coverage."),
+      createdAt: "2026-03-10T23:25:21Z",
+      updatedAt: "2026-03-10T23:25:21Z",
+      url: "https://example.test/issues/366",
+      state: "OPEN",
+    },
+  ];
+
+  let saveCalls = 0;
+  const stateStore = {
+    touch(record: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      return {
+        ...record,
+        ...patch,
+        updated_at: "2026-03-11T06:33:08.821Z",
+      };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  const recoveryEvents = await reconcileRecoverableBlockedIssueStates(stateStore, state, config, issues, {
+    shouldAutoRetryHandoffMissing,
+  });
+
+  const updated = state.issues["366"];
+  assert.equal(updated.state, "queued");
+  assert.equal(updated.blocked_reason, null);
+  assert.equal(updated.last_error, null);
+  assert.equal(updated.last_failure_context, null);
+  assert.equal(updated.last_failure_signature, null);
+  assert.equal(updated.repeated_failure_signature_count, 0);
+  assert.equal(updated.last_recovery_reason, "requirements_recovered: requeued issue #366 after execution-ready metadata was added");
+  assert.ok(updated.last_recovery_at);
+  assert.equal(saveCalls, 1);
+  assert.deepEqual(recoveryEvents.map((event) => event.reason), [
+    "requirements_recovered: requeued issue #366 after execution-ready metadata was added",
+  ]);
+});
+
+test("reconcileStaleActiveIssueReservation clears a stale reservation and emits a recovery loggable event", async () => {
+  const lockRoot = await fs.mkdtemp(path.join(os.tmpdir(), "codex-supervisor-locks-"));
+  const state: SupervisorStateFile = {
+    activeIssueNumber: 366,
+    issues: {
+      "366": createRecord({
+        issue_number: 366,
+        state: "implementing",
+        codex_session_id: "session-366",
+      }),
+    },
+  };
+
+  let saveCalls = 0;
+  const stateStore = {
+    touch(record: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      return {
+        ...record,
+        ...patch,
+        updated_at: "2026-03-11T06:33:08.821Z",
+      };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  const recoveryEvents = await reconcileStaleActiveIssueReservation({
+    stateStore,
+    state,
+    issueLockPath: (issueNumber) => path.join(lockRoot, "locks", "issues", String(issueNumber)),
+    sessionLockPath: (sessionId) => path.join(lockRoot, "locks", "sessions", String(sessionId)),
+  });
+
+  assert.equal(state.activeIssueNumber, null);
+  assert.equal(state.issues["366"]?.codex_session_id, null);
+  assert.equal(saveCalls, 1);
+  assert.equal(recoveryEvents.length, 1);
+  assert.match(
+    formatRecoveryLog(recoveryEvents) ?? "",
+    /recovery issue=#366 reason=stale_state_cleanup: cleared stale active reservation after issue lock and session lock were missing/,
+  );
+});
+
+test("reconcileMergedIssueClosures clears a stale active issue pointer even when the record already matches the done patch", async () => {
+  const original = createRecord({
+    issue_number: 366,
+    state: "done",
+    pr_number: null,
+    blocked_reason: null,
+    last_error: null,
+    last_failure_kind: null,
+    last_failure_context: null,
+    last_blocker_signature: null,
+    last_failure_signature: null,
+    timeout_retry_count: 0,
+    blocked_verification_retry_count: 0,
+    repeated_blocker_count: 0,
+    repeated_failure_signature_count: 0,
+    local_review_blocker_summary: null,
+    local_review_recommendation: null,
+    local_review_degraded: false,
+    external_review_head_sha: null,
+    external_review_misses_path: null,
+    external_review_matched_findings_count: 0,
+    external_review_near_match_findings_count: 0,
+    external_review_missed_findings_count: 0,
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: 366,
+    issues: {
+      "366": original,
+    },
+  };
+  const closedIssue: GitHubIssue = {
+    number: 366,
+    title: "Closed issue",
+    body: "",
+    createdAt: "2026-03-13T00:00:00Z",
+    updatedAt: "2026-03-13T00:20:00Z",
+    url: "https://example.test/issues/366",
+    state: "CLOSED",
+  };
+
+  let touchCalls = 0;
+  let saveCalls = 0;
+  const stateStore = {
+    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      touchCalls += 1;
+      return { ...current, ...patch };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  const recoveryEvents = await reconcileMergedIssueClosures(
+    {
+      getMergedPullRequestsClosingIssue: async () => [],
+      getPullRequestIfExists: async () => null,
+      closePullRequest: async () => {
+        throw new Error("unexpected closePullRequest call");
+      },
+      closeIssue: async () => {
+        throw new Error("unexpected closeIssue call");
+      },
+      getIssue: async () => {
+        throw new Error("unexpected getIssue call");
+      },
+      getChecks: async () => [],
+      getUnresolvedReviewThreads: async () => [],
+    },
+    stateStore,
+    state,
+    [closedIssue],
+  );
+
+  assert.equal(touchCalls, 0);
+  assert.equal(saveCalls, 1);
+  assert.equal(state.activeIssueNumber, null);
+  assert.deepEqual(recoveryEvents, []);
+  assert.deepEqual(state.issues["366"], original);
+});
+
+test("reconcileParentEpicClosures clears a stale active issue pointer even when the parent record already matches the done patch", async () => {
+  const original = createRecord({
+    issue_number: 123,
+    state: "done",
+    pr_number: null,
+    blocked_reason: null,
+    last_error: null,
+    last_failure_kind: null,
+    last_failure_context: null,
+    last_blocker_signature: null,
+    last_failure_signature: null,
+    timeout_retry_count: 0,
+    blocked_verification_retry_count: 0,
+    repeated_blocker_count: 0,
+    repeated_failure_signature_count: 0,
+    local_review_blocker_summary: null,
+    local_review_recommendation: null,
+    local_review_degraded: false,
+    external_review_head_sha: null,
+    external_review_misses_path: null,
+    external_review_matched_findings_count: 0,
+    external_review_near_match_findings_count: 0,
+    external_review_missed_findings_count: 0,
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: 123,
+    issues: {
+      "123": original,
+    },
+  };
+  const issues: GitHubIssue[] = [
+    {
+      number: 123,
+      title: "Parent issue",
+      body: "",
+      createdAt: "2026-03-13T00:00:00Z",
+      updatedAt: "2026-03-13T00:00:00Z",
+      url: "https://example.test/issues/123",
+      state: "OPEN",
+    },
+    {
+      number: 201,
+      title: "Child one",
+      body: "Part of #123",
+      createdAt: "2026-03-13T00:00:00Z",
+      updatedAt: "2026-03-13T00:00:00Z",
+      url: "https://example.test/issues/201",
+      state: "CLOSED",
+    },
+    {
+      number: 202,
+      title: "Child two",
+      body: "Part of: #123",
+      createdAt: "2026-03-13T00:00:00Z",
+      updatedAt: "2026-03-13T00:00:00Z",
+      url: "https://example.test/issues/202",
+      state: "CLOSED",
+    },
+  ];
+
+  let touchCalls = 0;
+  let saveCalls = 0;
+  let closeIssueCalls = 0;
+  const stateStore = {
+    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      touchCalls += 1;
+      return { ...current, ...patch };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  await reconcileParentEpicClosures(
+    {
+      closeIssue: async () => {
+        closeIssueCalls += 1;
+      },
+      closePullRequest: async () => {
+        throw new Error("unexpected closePullRequest call");
+      },
+      getChecks: async () => [],
+      getIssue: async () => {
+        throw new Error("unexpected getIssue call");
+      },
+      getMergedPullRequestsClosingIssue: async () => [],
+      getPullRequestIfExists: async () => null,
+      getUnresolvedReviewThreads: async () => [],
+    },
+    stateStore,
+    state,
+    issues,
+  );
+
+  assert.equal(closeIssueCalls, 1);
+  assert.equal(touchCalls, 0);
+  assert.equal(saveCalls, 1);
+  assert.equal(state.activeIssueNumber, null);
+  assert.deepEqual(state.issues["123"], original);
+});
+
+test("reconcileTrackedMergedButOpenIssues fetches missing issue snapshots for non-merging merged records", async () => {
+  const record = createRecord({
+    issue_number: 366,
+    state: "ready_to_merge",
+    pr_number: 191,
+    blocked_reason: null,
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: null,
+    issues: {
+      "366": record,
+    },
+  };
+  const mergedPr: GitHubPullRequest = {
+    number: 191,
+    title: "Merged implementation",
+    url: "https://example.test/pr/191",
+    state: "MERGED",
+    createdAt: "2026-03-13T00:10:00Z",
+    updatedAt: "2026-03-13T00:20:00Z",
+    isDraft: false,
+    headRefName: "codex/reopen-issue-366",
+    headRefOid: "merged-head-191",
+    mergeStateStatus: "CLEAN",
+    reviewDecision: "APPROVED",
+    mergedAt: "2026-03-13T00:20:00Z",
+    copilotReviewState: null,
+    copilotReviewRequestedAt: null,
+    copilotReviewArrivedAt: null,
+  };
+  const closedIssue: GitHubIssue = {
+    number: 366,
+    title: "Merged implementation issue",
+    body: "",
+    createdAt: "2026-03-13T00:00:00Z",
+    updatedAt: "2026-03-13T00:21:00Z",
+    url: "https://example.test/issues/366",
+    state: "CLOSED",
+  };
+
+  let getIssueCalls = 0;
+  let saveCalls = 0;
+  const stateStore = {
+    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      return {
+        ...current,
+        ...patch,
+        updated_at: "2026-03-13T00:25:00Z",
+      };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  const recoveryEvents = await reconcileTrackedMergedButOpenIssues(
+    {
+      getPullRequestIfExists: async () => mergedPr,
+      getIssue: async () => {
+        getIssueCalls += 1;
+        return closedIssue;
+      },
+      closeIssue: async () => {
+        throw new Error("unexpected closeIssue call");
+      },
+      closePullRequest: async () => {
+        throw new Error("unexpected closePullRequest call");
+      },
+      getChecks: async () => [],
+      getMergedPullRequestsClosingIssue: async () => [],
+      getUnresolvedReviewThreads: async () => [],
+    },
+    stateStore,
+    state,
+    [],
+  );
+
+  assert.equal(getIssueCalls, 1);
+  assert.equal(saveCalls, 1);
+  assert.equal(state.issues["366"]?.state, "done");
+  assert.equal(state.issues["366"]?.pr_number, 191);
+  assert.equal(state.issues["366"]?.last_head_sha, "merged-head-191");
+  assert.deepEqual(recoveryEvents.map((event) => event.reason), [
+    "merged_pr_convergence: tracked PR #191 merged; marked issue #366 done",
+  ]);
+});
+
+test("reconcileTrackedMergedButOpenIssues does not rewrite recovery metadata when the done state is already current", async () => {
+  const original = createRecord({
+    issue_number: 366,
+    state: "done",
+    pr_number: 191,
+    blocked_reason: null,
+    last_error: null,
+    last_failure_kind: null,
+    last_failure_context: null,
+    last_blocker_signature: null,
+    last_failure_signature: null,
+    timeout_retry_count: 0,
+    blocked_verification_retry_count: 0,
+    repeated_blocker_count: 0,
+    repeated_failure_signature_count: 0,
+    last_head_sha: "merged-head-191",
+    local_review_blocker_summary: null,
+    local_review_recommendation: null,
+    local_review_degraded: false,
+    external_review_head_sha: null,
+    external_review_misses_path: null,
+    external_review_matched_findings_count: 0,
+    external_review_near_match_findings_count: 0,
+    external_review_missed_findings_count: 0,
+    last_recovery_reason: "existing recovery reason",
+    last_recovery_at: "2026-03-13T00:30:00Z",
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: null,
+    issues: {
+      "366": original,
+    },
+  };
+  const mergedPr: GitHubPullRequest = {
+    number: 191,
+    title: "Merged implementation",
+    url: "https://example.test/pr/191",
+    state: "MERGED",
+    createdAt: "2026-03-13T00:10:00Z",
+    updatedAt: "2026-03-13T00:20:00Z",
+    isDraft: false,
+    headRefName: "codex/reopen-issue-366",
+    headRefOid: "merged-head-191",
+    mergeStateStatus: "CLEAN",
+    reviewDecision: "APPROVED",
+    mergedAt: "2026-03-13T00:20:00Z",
+    copilotReviewState: null,
+    copilotReviewRequestedAt: null,
+    copilotReviewArrivedAt: null,
+  };
+  const closedIssue: GitHubIssue = {
+    number: 366,
+    title: "Merged implementation issue",
+    body: "",
+    createdAt: "2026-03-13T00:00:00Z",
+    updatedAt: "2026-03-13T00:21:00Z",
+    url: "https://example.test/issues/366",
+    state: "CLOSED",
+  };
+
+  let touchCalls = 0;
+  let saveCalls = 0;
+  const stateStore = {
+    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      touchCalls += 1;
+      return {
+        ...current,
+        ...patch,
+        updated_at: "2026-03-13T00:35:00Z",
+      };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  const recoveryEvents = await reconcileTrackedMergedButOpenIssues(
+    {
+      getPullRequestIfExists: async () => mergedPr,
+      getIssue: async () => closedIssue,
+      closeIssue: async () => {
+        throw new Error("unexpected closeIssue call");
+      },
+      closePullRequest: async () => {
+        throw new Error("unexpected closePullRequest call");
+      },
+      getChecks: async () => [],
+      getMergedPullRequestsClosingIssue: async () => [],
+      getUnresolvedReviewThreads: async () => [],
+    },
+    stateStore,
+    state,
+    [closedIssue],
+  );
+
+  assert.equal(touchCalls, 0);
+  assert.equal(saveCalls, 0);
+  assert.deepEqual(recoveryEvents, []);
+  assert.deepEqual(state.issues["366"], original);
+});
+
+test("runOnce recovers when post-codex refresh throws after leaving a dirty worktree", async () => {
+  const fixture = await createSupervisorFixture();
+  const issueNumber = 87;
+  const branch = branchName(fixture.config, issueNumber);
+  const state: SupervisorStateFile = {
+    activeIssueNumber: issueNumber,
+    issues: {
+      [String(issueNumber)]: createRecord({
+        issue_number: issueNumber,
+        state: "stabilizing",
+        branch,
+        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
+        journal_path: null,
+        attempt_count: 1,
+        implementation_attempt_count: 1,
+        repair_attempt_count: 0,
+        last_error: null,
+        last_failure_kind: null,
+        last_failure_context: null,
+        last_failure_signature: null,
+        repeated_failure_signature_count: 0,
+        blocked_reason: null,
+      }),
+    },
+  };
+  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
+
+  const issue: GitHubIssue = {
+    number: issueNumber,
+    title: "Reproduce dirty worktree recovery",
+    body: executionReadyBody("Reproduce dirty worktree recovery."),
+    createdAt: "2026-03-13T00:00:00Z",
+    updatedAt: "2026-03-13T00:00:00Z",
+    url: `https://example.test/issues/${issueNumber}`,
+    state: "OPEN",
+  };
+
+  let resolveCalls = 0;
+  const supervisor = new Supervisor(fixture.config);
+  (supervisor as unknown as { github: Record<string, unknown> }).github = {
+    authStatus: async () => ({ ok: true, message: null }),
+    listAllIssues: async () => [issue],
+    listCandidateIssues: async () => [issue],
+    getIssue: async () => issue,
+    resolvePullRequestForBranch: async () => {
+      resolveCalls += 1;
+      if (resolveCalls === 1) {
+        return null;
+      }
+      throw new Error("post-turn refresh blew up");
+    },
+    getChecks: async () => [],
+    getUnresolvedReviewThreads: async () => [],
+    getPullRequestIfExists: async () => null,
+    getMergedPullRequestsClosingIssue: async () => [],
+    closeIssue: async () => {
+      throw new Error("unexpected closeIssue call");
+    },
+  };
+
+  const message = await supervisor.runOnce({ dryRun: false });
+  assert.match(message, /Recovered from unexpected Codex turn failure/);
+
+  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
+  const record = persisted.issues[String(issueNumber)];
+  assert.equal(persisted.activeIssueNumber, null);
+  assert.equal(record.state, "failed");
+  assert.equal(record.last_failure_kind, "command_error");
+  assert.match(record.last_error ?? "", /post-turn refresh blew up/);
+  assert.equal(record.codex_session_id, "thread-123");
+  assert.match(record.last_codex_summary ?? "", /created a dirty checkpoint/);
+  assert.equal(record.blocked_reason, null);
+  assert.match(record.last_failure_context?.summary ?? "", /Supervisor failed while recovering a Codex turn/);
+  assert.deepEqual(record.last_failure_context?.details.slice(0, 4), [
+    "previous_state=stabilizing",
+    "workspace_dirty=yes",
+    `workspace_head=${record.last_head_sha}`,
+    "pr_number=none",
+  ]);
+
+  const worktreeStatus = git(["-C", path.join(fixture.workspaceRoot, `issue-${issueNumber}`), "status", "--short"]);
+  assert.match(worktreeStatus, /dirty\.txt/);
+
+  const issueLockPath = path.join(path.dirname(fixture.stateFile), "locks", "issues", `issue-${issueNumber}.lock`);
+  await assert.rejects(fs.access(issueLockPath));
+});
+
+test("recoverUnexpectedCodexTurnFailure preserves dirty recovery context and timeout bookkeeping", async () => {
+  const issueNumber = 88;
+  const record = createRecord({
+    issue_number: issueNumber,
+    state: "stabilizing",
+    timeout_retry_count: 1,
+    codex_session_id: "thread-456",
+    last_head_sha: "abc1234",
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: issueNumber,
+    issues: {
+      [String(issueNumber)]: record,
+    },
+  };
+  let saveCalls = 0;
+  let syncedRecord: IssueRunRecord | null = null;
+  const stateStore = {
+    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      return { ...current, ...patch, updated_at: "2026-03-13T00:00:00.000Z" };
+    },
+    async save(): Promise<void> {
+      saveCalls += 1;
+    },
+  };
+
+  const updated = await recoverUnexpectedCodexTurnFailure({
+    stateStore: stateStore as unknown as Parameters<typeof recoverUnexpectedCodexTurnFailure>[0]["stateStore"],
+    state,
+    record,
+    issue: {
+      number: issueNumber,
+      title: "Timeout while recovering dirty worktree",
+      body: "",
+      createdAt: "2026-03-13T00:00:00Z",
+      updatedAt: "2026-03-13T00:00:00Z",
+      url: `https://example.test/issues/${issueNumber}`,
+      state: "OPEN",
+    },
+    journalSync: async (nextRecord) => {
+      syncedRecord = nextRecord;
+    },
+    error: new Error("Command timed out after 1800000ms: codex exec resume thread-456"),
+    workspaceStatus: {
+      hasUncommittedChanges: true,
+      headSha: "deadbee",
+    },
+    pr: {
+      number: 55,
+      headRefOid: "feed123",
+    },
+  });
+
+  assert.equal(saveCalls, 1);
+  assert.equal(state.activeIssueNumber, null);
+  assert.equal(updated.state, "failed");
+  assert.equal(updated.last_failure_kind, "timeout");
+  assert.equal(updated.timeout_retry_count, 2);
+  assert.equal(updated.blocked_reason, null);
+  assert.match(updated.last_error ?? "", /Command timed out after 1800000ms/);
+  assert.match(updated.last_failure_context?.summary ?? "", /Supervisor failed while recovering a Codex turn/);
+  assert.deepEqual(updated.last_failure_context?.details.slice(0, 6), [
+    "previous_state=stabilizing",
+    "workspace_dirty=yes",
+    "workspace_head=deadbee",
+    "pr_number=55",
+    "pr_head=feed123",
+    "codex_session_id=thread-456",
+  ]);
+  assert.equal(state.issues[String(issueNumber)], updated);
+  assert.equal(syncedRecord, updated);
+});
+
+test("recoverUnexpectedCodexTurnFailure records unavailable workspace inspection distinctly", async () => {
+  const issueNumber = 92;
+  const record = createRecord({
+    issue_number: issueNumber,
+    state: "stabilizing",
+    last_head_sha: "abc1234",
+    codex_session_id: null,
+  });
+  const state: SupervisorStateFile = {
+    activeIssueNumber: issueNumber,
+    issues: {
+      [String(issueNumber)]: record,
+    },
+  };
+  const stateStore = {
+    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
+      return { ...current, ...patch, updated_at: "2026-03-15T02:00:00.000Z" };
+    },
+    async save(): Promise<void> {},
+  };
+
+  const updated = await recoverUnexpectedCodexTurnFailure({
+    stateStore: stateStore as unknown as Parameters<typeof recoverUnexpectedCodexTurnFailure>[0]["stateStore"],
+    state,
+    record,
+    issue: {
+      number: issueNumber,
+      title: "Refresh failed before workspace inspection",
+      body: "",
+      createdAt: "2026-03-15T00:00:00Z",
+      updatedAt: "2026-03-15T00:00:00Z",
+      url: `https://example.test/issues/${issueNumber}`,
+      state: "OPEN",
+    },
+    journalSync: async () => {},
+    error: new Error("Command failed: codex exec resume"),
+    workspaceStatus: null,
+    pr: null,
+  });
+
+  assert.deepEqual(updated.last_failure_context?.details.slice(0, 6), [
+    "previous_state=stabilizing",
+    "workspace_dirty=unknown",
+    "workspace_head=abc1234",
+    "pr_number=none",
+    "pr_head=none",
+    "codex_session_id=none",
+  ]);
+});
