+++ src/supervisor/supervisor.test.ts
@@ -1,191 +1,24 @@
 import test from "node:test";
 import assert from "node:assert/strict";
-import { execFileSync } from "node:child_process";
 import fs from "node:fs/promises";
-import os from "node:os";
 import path from "node:path";
 import {
   Supervisor,
 } from "./supervisor";
 import { AgentRunner, AgentTurnRequest } from "./agent-runner";
-import { recoverUnexpectedCodexTurnFailure } from "./supervisor-failure-helpers";
 import { formatDetailedStatus } from "./supervisor-status-rendering";
-import { shouldAutoRetryHandoffMissing } from "./supervisor-execution-policy";
 import { handleAuthFailure } from "./supervisor-failure-helpers";
-import {
-  formatRecoveryLog,
-  reconcileMergedIssueClosures,
-  reconcileParentEpicClosures,
-  reconcileRecoverableBlockedIssueStates,
-  reconcileStaleActiveIssueReservation,
-  reconcileTrackedMergedButOpenIssues,
-} from "../recovery-reconciliation";
 import { StateStore } from "../core/state-store";
 import { GitHubIssue, GitHubPullRequest, IssueRunRecord, PullRequestCheck, ReviewThread, SupervisorConfig, SupervisorStateFile } from "../core/types";
-
-function createConfig(overrides: Partial<SupervisorConfig> = {}): SupervisorConfig {
-  return {
-    repoPath: "/tmp/repo",
-    repoSlug: "owner/repo",
-    defaultBranch: "main",
-    workspaceRoot: "/tmp/workspaces",
-    stateBackend: "json",
-    stateFile: "/tmp/state.json",
-    codexBinary: "/usr/bin/codex",
-    codexModelStrategy: "inherit",
-    codexReasoningEffortByState: {},
-    codexReasoningEscalateOnRepeatedFailure: true,
-    sharedMemoryFiles: [],
-    gsdEnabled: false,
-    gsdAutoInstall: false,
-    gsdInstallScope: "global",
-    gsdPlanningFiles: [],
-    localReviewEnabled: false,
-    localReviewAutoDetect: true,
-    localReviewRoles: [],
-    localReviewArtifactDir: "/tmp/reviews",
-    localReviewConfidenceThreshold: 0.7,
-    localReviewReviewerThresholds: {
-      generic: {
-        confidenceThreshold: 0.7,
-        minimumSeverity: "low",
-      },
-      specialist: {
-        confidenceThreshold: 0.7,
-        minimumSeverity: "low",
-      },
-    },
-    localReviewPolicy: "block_ready",
-    localReviewHighSeverityAction: "retry",
-    reviewBotLogins: [],
-    humanReviewBlocksMerge: true,
-    issueJournalRelativePath: ".codex-supervisor/issue-journal.md",
-    issueJournalMaxChars: 6000,
-    skipTitlePrefixes: [],
-    branchPrefix: "codex/reopen-issue-",
-    pollIntervalSeconds: 60,
-    copilotReviewWaitMinutes: 10,
-    copilotReviewTimeoutAction: "continue",
-    codexExecTimeoutMinutes: 30,
-    maxCodexAttemptsPerIssue: 5,
-    maxImplementationAttemptsPerIssue: 5,
-    maxRepairAttemptsPerIssue: 5,
-    timeoutRetryLimit: 2,
-    blockedVerificationRetryLimit: 3,
-    sameBlockerRepeatLimit: 2,
-    sameFailureSignatureRepeatLimit: 3,
-    maxDoneWorkspaces: 24,
-    cleanupDoneWorkspacesAfterHours: 24,
-    mergeMethod: "squash",
-    draftPrAfterAttempt: 1,
-    ...overrides,
-  };
-}
-
-function createRecord(overrides: Partial<IssueRunRecord> = {}): IssueRunRecord {
-  return {
-    issue_number: 366,
-    state: "blocked",
-    branch: "codex/reopen-issue-366",
-    pr_number: null,
-    workspace: "/tmp/workspaces/issue-366",
-    journal_path: "/tmp/workspaces/issue-366/.codex-supervisor/issue-journal.md",
-    review_wait_started_at: null,
-    review_wait_head_sha: null,
-    copilot_review_requested_observed_at: null,
-    copilot_review_requested_head_sha: null,
-    copilot_review_timed_out_at: null,
-    copilot_review_timeout_action: null,
-    copilot_review_timeout_reason: null,
-    codex_session_id: "session-1",
-    local_review_head_sha: null,
-    local_review_blocker_summary: null,
-    local_review_summary_path: null,
-    local_review_run_at: null,
-    local_review_max_severity: null,
-    local_review_findings_count: 0,
-    local_review_root_cause_count: 0,
-    local_review_verified_max_severity: null,
-    local_review_verified_findings_count: 0,
-    local_review_recommendation: null,
-    local_review_degraded: false,
-    last_local_review_signature: null,
-    repeated_local_review_signature_count: 0,
-    external_review_head_sha: null,
-    external_review_misses_path: null,
-    external_review_matched_findings_count: 0,
-    external_review_near_match_findings_count: 0,
-    external_review_missed_findings_count: 0,
-    attempt_count: 2,
-    implementation_attempt_count: 2,
-    repair_attempt_count: 0,
-    timeout_retry_count: 0,
-    blocked_verification_retry_count: 0,
-    repeated_blocker_count: 0,
-    repeated_failure_signature_count: 1,
-    last_head_sha: "abcdef1",
-    last_codex_summary: null,
-    last_recovery_reason: null,
-    last_recovery_at: null,
-    last_error: "Codex completed without updating the issue journal for issue #366.",
-    last_failure_kind: null,
-    last_failure_context: {
-      category: "blocked",
-      summary: "Codex completed without updating the issue journal for issue #366.",
-      signature: "handoff-missing",
-      command: null,
-      details: ["Update the Codex Working Notes section before ending the turn."],
-      url: null,
-      updated_at: "2026-03-11T01:50:41.997Z",
-    },
-    last_blocker_signature: null,
-    last_failure_signature: "handoff-missing",
-    blocked_reason: "handoff_missing",
-    processed_review_thread_ids: [],
-    processed_review_thread_fingerprints: [],
-    updated_at: "2026-03-11T01:50:41.997Z",
-    ...overrides,
-  };
-}
-
-function createReviewThread(overrides: Partial<ReviewThread> = {}): ReviewThread {
-  return {
-    id: "thread-1",
-    isResolved: false,
-    isOutdated: false,
-    path: "src/file.ts",
-    line: 12,
-    comments: {
-      nodes: [
-        {
-          id: "comment-1",
-          body: "Please address this.",
-          createdAt: "2026-03-11T00:00:00Z",
-          url: "https://example.test/pr/44#discussion_r1",
-          author: {
-            login: "copilot-pull-request-reviewer",
-            typeName: "Bot",
-          },
-        },
-      ],
-    },
-    ...overrides,
-  };
-}
-
-function executionReadyBody(summary: string): string {
-  return `## Summary
-${summary}
-
-## Scope
-- keep the test fixture execution-ready
-
-## Acceptance criteria
-- supervisor treats this issue as runnable
-
-## Verification
-- npm test -- src/supervisor.test.ts`;
-}
+import {
+  branchName,
+  createConfig,
+  createRecord,
+  createReviewThread,
+  createSupervisorFixture,
+  executionReadyBody,
+  git,
+} from "./supervisor-test-helpers";
 
 function withStubbedDateNow<T>(nowIso: string, run: () => T): T {
   const originalDateNow = Date.now;
@@ -196,818 +29,6 @@ function withStubbedDateNow<T>(nowIso: string, run: () => T): T {
     Date.now = originalDateNow;
   }
 }
-
-function git(args: string[], cwd?: string): string {
-  return execFileSync("git", args, {
-    cwd,
-    encoding: "utf8",
-    env: {
-      ...process.env,
-      GIT_AUTHOR_NAME: "Codex",
-      GIT_AUTHOR_EMAIL: "codex@example.com",
-      GIT_COMMITTER_NAME: "Codex",
-      GIT_COMMITTER_EMAIL: "codex@example.com",
-    },
-  }).trim();
-}
-
-async function createSupervisorFixture(options: {
-  codexScriptLines?: string[];
-} = {}): Promise<{
-  config: SupervisorConfig;
-  repoPath: string;
-  stateFile: string;
-  workspaceRoot: string;
-}> {
-  const root = await fs.mkdtemp(path.join(os.tmpdir(), "codex-supervisor-issue-87-"));
-  const remotePath = path.join(root, "remote.git");
-  const seedPath = path.join(root, "seed");
-  const repoPath = path.join(root, "repo");
-  const workspaceRoot = path.join(root, "workspaces");
-  const stateFile = path.join(root, "state.json");
-  const codexBinary = path.join(root, "fake-codex.sh");
-
-  git(["init", "--bare", remotePath]);
-  await fs.mkdir(seedPath, { recursive: true });
-  git(["init", "-b", "main"], seedPath);
-  await fs.writeFile(path.join(seedPath, "README.md"), "# fixture\n", "utf8");
-  git(["add", "README.md"], seedPath);
-  git(["commit", "-m", "seed"], seedPath);
-  git(["remote", "add", "origin", remotePath], seedPath);
-  git(["push", "-u", "origin", "main"], seedPath);
-  git(["symbolic-ref", "HEAD", "refs/heads/main"], remotePath);
-  git(["clone", remotePath, repoPath]);
-  git(["-C", repoPath, "branch", "--set-upstream-to=origin/main", "main"]);
-
-  const codexScriptLines = options.codexScriptLines ?? [
-    "#!/bin/sh",
-    "set -eu",
-    'out=""',
-    'while [ "$#" -gt 0 ]; do',
-    '  case "$1" in',
-    '    -o) out="$2"; shift 2 ;;',
-    '    *) shift ;;',
-    '  esac',
-    "done",
-    'printf \'{"type":"thread.started","thread_id":"thread-123"}\\n\'',
-    "cat <<'EOF' > \"$out\"",
-    "Summary: created a dirty checkpoint",
-    "State hint: stabilizing",
-    "Blocked reason: none",
-    "Tests: not run",
-    "Failure signature: none",
-    "Next action: inspect the dirty worktree and finish recovery",
-    "EOF",
-    "printf '\\n- Scratchpad note: codex wrote a dirty change for reproduction.\\n' >> .codex-supervisor/issue-journal.md",
-    "printf 'dirty change\\n' >> dirty.txt",
-    "exit 0",
-    "",
-  ];
-  await fs.writeFile(codexBinary, codexScriptLines.join("\n"), "utf8");
-  await fs.chmod(codexBinary, 0o755);
-
-  return {
-    repoPath,
-    stateFile,
-    workspaceRoot,
-    config: createConfig({
-      repoPath,
-      workspaceRoot,
-      stateFile,
-      codexBinary,
-      issueJournalMaxChars: 12000,
-    }),
-  };
-}
-
-test("doctor uses the diagnostic-only state loader instead of StateStore.load", async (t) => {
-  const fixture = await createSupervisorFixture();
-  t.after(async () => {
-    await fs.rm(path.dirname(fixture.repoPath), { recursive: true, force: true });
-  });
-
-  await fs.writeFile(fixture.stateFile, "{not-json}\n", "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-  };
-
-  const stateStore = (supervisor as unknown as { stateStore: StateStore }).stateStore;
-  stateStore.load = async () => {
-    throw new Error("StateStore.load should not be used by doctor");
-  };
-
-  const report = await supervisor.doctor();
-
-  assert.match(report, /doctor_check name=github_auth status=pass/);
-  assert.match(report, /doctor_check name=state_file status=fail/);
-  assert.match(report, /doctor_check name=worktrees status=fail/);
-});
-
-test("shouldAutoRetryHandoffMissing only retries recoverable blocked handoffs", () => {
-  const config = createConfig({
-    maxImplementationAttemptsPerIssue: 3,
-    maxRepairAttemptsPerIssue: 9,
-  });
-
-  assert.equal(shouldAutoRetryHandoffMissing(createRecord(), config), true);
-  assert.equal(
-    shouldAutoRetryHandoffMissing(
-      createRecord({
-        attempt_count: 8,
-        implementation_attempt_count: 2,
-        repair_attempt_count: 6,
-      }),
-      config,
-    ),
-    true,
-  );
-  assert.equal(shouldAutoRetryHandoffMissing(createRecord({ pr_number: 12 }), config), false);
-  assert.equal(
-    shouldAutoRetryHandoffMissing(
-      createRecord({ repeated_failure_signature_count: config.sameFailureSignatureRepeatLimit }),
-      config,
-    ),
-    false,
-  );
-  assert.equal(
-    shouldAutoRetryHandoffMissing(
-      createRecord({
-        attempt_count: config.maxImplementationAttemptsPerIssue + 6,
-        implementation_attempt_count: config.maxImplementationAttemptsPerIssue,
-        repair_attempt_count: 6,
-      }),
-      config,
-    ),
-    false,
-  );
-});
-
-test("reconcileRecoverableBlockedIssueStates requeues open handoff-missing issues without dropping repeat tracking", async () => {
-  const config = createConfig();
-  const original = createRecord();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "366": original,
-    },
-  };
-  const issues: GitHubIssue[] = [
-    {
-      number: 366,
-      title: "P3: Add regression coverage",
-      body: "",
-      createdAt: "2026-03-10T23:25:21Z",
-      updatedAt: "2026-03-10T23:25:21Z",
-      url: "https://example.test/issues/366",
-      state: "OPEN",
-    },
-  ];
-
-  let saveCalls = 0;
-  const stateStore = {
-    touch(record: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      return {
-        ...record,
-        ...patch,
-        updated_at: "2026-03-11T06:33:08.821Z",
-      };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  await reconcileRecoverableBlockedIssueStates(stateStore, state, config, issues, {
-    shouldAutoRetryHandoffMissing,
-  });
-
-  const updated = state.issues["366"];
-  assert.equal(updated.state, "queued");
-  assert.equal(updated.blocked_reason, null);
-  assert.equal(updated.last_error, null);
-  assert.equal(updated.codex_session_id, null);
-  assert.equal(updated.last_failure_signature, "handoff-missing");
-  assert.equal(
-    updated.last_failure_context?.summary ?? null,
-    "Codex completed without updating the issue journal for issue #366.",
-  );
-  assert.equal(updated.repeated_failure_signature_count, 1);
-  assert.equal(saveCalls, 1);
-});
-
-test("reconcileRecoverableBlockedIssueStates leaves closed issues blocked", async () => {
-  const config = createConfig();
-  const original = createRecord();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "366": original,
-    },
-  };
-  const issues: GitHubIssue[] = [
-    {
-      number: 366,
-      title: "P3: Add regression coverage",
-      body: "",
-      createdAt: "2026-03-10T23:25:21Z",
-      updatedAt: "2026-03-10T23:25:21Z",
-      url: "https://example.test/issues/366",
-      state: "CLOSED",
-    },
-  ];
-
-  let saveCalls = 0;
-  const stateStore = {
-    touch(record: IssueRunRecord): IssueRunRecord {
-      return record;
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  await reconcileRecoverableBlockedIssueStates(stateStore, state, config, issues, {
-    shouldAutoRetryHandoffMissing,
-  });
-
-  assert.deepEqual(state.issues["366"], original);
-  assert.equal(saveCalls, 0);
-});
-
-test("reconcileRecoverableBlockedIssueStates requeues requirements-blocked issues once metadata is execution-ready", async () => {
-  const config = createConfig();
-  const original = createRecord({
-    state: "blocked",
-    blocked_reason: "requirements",
-    last_error: "Missing required execution-ready metadata: scope, acceptance criteria, verification.",
-    last_failure_kind: null,
-    last_failure_context: {
-      category: "blocked",
-      summary: "Issue #366 is not execution-ready because it is missing: scope, acceptance criteria, verification.",
-      signature: "requirements:scope|acceptance criteria|verification",
-      command: null,
-      details: [
-        "missing_required=scope, acceptance criteria, verification",
-        "missing_recommended=depends on, execution order",
-      ],
-      url: "https://example.test/issues/366",
-      updated_at: "2026-03-11T01:50:41.997Z",
-    },
-    last_failure_signature: "requirements:scope|acceptance criteria|verification",
-    repeated_failure_signature_count: 2,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "366": original,
-    },
-  };
-  const issues: GitHubIssue[] = [
-    {
-      number: 366,
-      title: "P3: Add regression coverage",
-      body: executionReadyBody("Add regression coverage."),
-      createdAt: "2026-03-10T23:25:21Z",
-      updatedAt: "2026-03-10T23:25:21Z",
-      url: "https://example.test/issues/366",
-      state: "OPEN",
-    },
-  ];
-
-  let saveCalls = 0;
-  const stateStore = {
-    touch(record: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      return {
-        ...record,
-        ...patch,
-        updated_at: "2026-03-11T06:33:08.821Z",
-      };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  const recoveryEvents = await reconcileRecoverableBlockedIssueStates(stateStore, state, config, issues, {
-    shouldAutoRetryHandoffMissing,
-  });
-
-  const updated = state.issues["366"];
-  assert.equal(updated.state, "queued");
-  assert.equal(updated.blocked_reason, null);
-  assert.equal(updated.last_error, null);
-  assert.equal(updated.last_failure_context, null);
-  assert.equal(updated.last_failure_signature, null);
-  assert.equal(updated.repeated_failure_signature_count, 0);
-  assert.equal(updated.last_recovery_reason, "requirements_recovered: requeued issue #366 after execution-ready metadata was added");
-  assert.ok(updated.last_recovery_at);
-  assert.equal(saveCalls, 1);
-  assert.deepEqual(recoveryEvents.map((event) => event.reason), [
-    "requirements_recovered: requeued issue #366 after execution-ready metadata was added",
-  ]);
-});
-
-test("reconcileStaleActiveIssueReservation clears a stale reservation and emits a recovery loggable event", async () => {
-  const lockRoot = await fs.mkdtemp(path.join(os.tmpdir(), "codex-supervisor-locks-"));
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 366,
-    issues: {
-      "366": createRecord({
-        issue_number: 366,
-        state: "implementing",
-        codex_session_id: "session-366",
-      }),
-    },
-  };
-
-  let saveCalls = 0;
-  const stateStore = {
-    touch(record: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      return {
-        ...record,
-        ...patch,
-        updated_at: "2026-03-11T06:33:08.821Z",
-      };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  const recoveryEvents = await reconcileStaleActiveIssueReservation({
-    stateStore,
-    state,
-    issueLockPath: (issueNumber) => path.join(lockRoot, "locks", "issues", String(issueNumber)),
-    sessionLockPath: (sessionId) => path.join(lockRoot, "locks", "sessions", String(sessionId)),
-  });
-
-  assert.equal(state.activeIssueNumber, null);
-  assert.equal(state.issues["366"]?.codex_session_id, null);
-  assert.equal(saveCalls, 1);
-  assert.equal(recoveryEvents.length, 1);
-  assert.match(
-    formatRecoveryLog(recoveryEvents) ?? "",
-    /recovery issue=#366 reason=stale_state_cleanup: cleared stale active reservation after issue lock and session lock were missing/,
-  );
-});
-
-test("reconcileMergedIssueClosures clears a stale active issue pointer even when the record already matches the done patch", async () => {
-  const original = createRecord({
-    issue_number: 366,
-    state: "done",
-    pr_number: null,
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_blocker_signature: null,
-    last_failure_signature: null,
-    timeout_retry_count: 0,
-    blocked_verification_retry_count: 0,
-    repeated_blocker_count: 0,
-    repeated_failure_signature_count: 0,
-    local_review_blocker_summary: null,
-    local_review_recommendation: null,
-    local_review_degraded: false,
-    external_review_head_sha: null,
-    external_review_misses_path: null,
-    external_review_matched_findings_count: 0,
-    external_review_near_match_findings_count: 0,
-    external_review_missed_findings_count: 0,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 366,
-    issues: {
-      "366": original,
-    },
-  };
-  const closedIssue: GitHubIssue = {
-    number: 366,
-    title: "Closed issue",
-    body: "",
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:20:00Z",
-    url: "https://example.test/issues/366",
-    state: "CLOSED",
-  };
-
-  let touchCalls = 0;
-  let saveCalls = 0;
-  const stateStore = {
-    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      touchCalls += 1;
-      return { ...current, ...patch };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  const recoveryEvents = await reconcileMergedIssueClosures(
-    {
-      getMergedPullRequestsClosingIssue: async () => [],
-      getPullRequestIfExists: async () => null,
-      closePullRequest: async () => {
-        throw new Error("unexpected closePullRequest call");
-      },
-      closeIssue: async () => {
-        throw new Error("unexpected closeIssue call");
-      },
-      getIssue: async () => {
-        throw new Error("unexpected getIssue call");
-      },
-      getChecks: async () => [],
-      getUnresolvedReviewThreads: async () => [],
-    },
-    stateStore,
-    state,
-    [closedIssue],
-  );
-
-  assert.equal(touchCalls, 0);
-  assert.equal(saveCalls, 1);
-  assert.equal(state.activeIssueNumber, null);
-  assert.deepEqual(recoveryEvents, []);
-  assert.deepEqual(state.issues["366"], original);
-});
-
-test("reconcileParentEpicClosures clears a stale active issue pointer even when the parent record already matches the done patch", async () => {
-  const original = createRecord({
-    issue_number: 123,
-    state: "done",
-    pr_number: null,
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_blocker_signature: null,
-    last_failure_signature: null,
-    timeout_retry_count: 0,
-    blocked_verification_retry_count: 0,
-    repeated_blocker_count: 0,
-    repeated_failure_signature_count: 0,
-    local_review_blocker_summary: null,
-    local_review_recommendation: null,
-    local_review_degraded: false,
-    external_review_head_sha: null,
-    external_review_misses_path: null,
-    external_review_matched_findings_count: 0,
-    external_review_near_match_findings_count: 0,
-    external_review_missed_findings_count: 0,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 123,
-    issues: {
-      "123": original,
-    },
-  };
-  const issues: GitHubIssue[] = [
-    {
-      number: 123,
-      title: "Parent issue",
-      body: "",
-      createdAt: "2026-03-13T00:00:00Z",
-      updatedAt: "2026-03-13T00:00:00Z",
-      url: "https://example.test/issues/123",
-      state: "OPEN",
-    },
-    {
-      number: 201,
-      title: "Child one",
-      body: "Part of #123",
-      createdAt: "2026-03-13T00:00:00Z",
-      updatedAt: "2026-03-13T00:00:00Z",
-      url: "https://example.test/issues/201",
-      state: "CLOSED",
-    },
-    {
-      number: 202,
-      title: "Child two",
-      body: "Part of: #123",
-      createdAt: "2026-03-13T00:00:00Z",
-      updatedAt: "2026-03-13T00:00:00Z",
-      url: "https://example.test/issues/202",
-      state: "CLOSED",
-    },
-  ];
-
-  let touchCalls = 0;
-  let saveCalls = 0;
-  let closeIssueCalls = 0;
-  const stateStore = {
-    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      touchCalls += 1;
-      return { ...current, ...patch };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  await reconcileParentEpicClosures(
-    {
-      closeIssue: async () => {
-        closeIssueCalls += 1;
-      },
-      closePullRequest: async () => {
-        throw new Error("unexpected closePullRequest call");
-      },
-      getChecks: async () => [],
-      getIssue: async () => {
-        throw new Error("unexpected getIssue call");
-      },
-      getMergedPullRequestsClosingIssue: async () => [],
-      getPullRequestIfExists: async () => null,
-      getUnresolvedReviewThreads: async () => [],
-    },
-    stateStore,
-    state,
-    issues,
-  );
-
-  assert.equal(closeIssueCalls, 1);
-  assert.equal(touchCalls, 0);
-  assert.equal(saveCalls, 1);
-  assert.equal(state.activeIssueNumber, null);
-  assert.deepEqual(state.issues["123"], original);
-});
-
-test("reconcileTrackedMergedButOpenIssues fetches missing issue snapshots for non-merging merged records", async () => {
-  const record = createRecord({
-    issue_number: 366,
-    state: "ready_to_merge",
-    pr_number: 191,
-    blocked_reason: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "366": record,
-    },
-  };
-  const mergedPr: GitHubPullRequest = {
-    number: 191,
-    title: "Merged implementation",
-    url: "https://example.test/pr/191",
-    state: "MERGED",
-    createdAt: "2026-03-13T00:10:00Z",
-    updatedAt: "2026-03-13T00:20:00Z",
-    isDraft: false,
-    headRefName: "codex/reopen-issue-366",
-    headRefOid: "merged-head-191",
-    mergeStateStatus: "CLEAN",
-    reviewDecision: "APPROVED",
-    mergedAt: "2026-03-13T00:20:00Z",
-    copilotReviewState: null,
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-  const closedIssue: GitHubIssue = {
-    number: 366,
-    title: "Merged implementation issue",
-    body: "",
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:21:00Z",
-    url: "https://example.test/issues/366",
-    state: "CLOSED",
-  };
-
-  let getIssueCalls = 0;
-  let saveCalls = 0;
-  const stateStore = {
-    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      return {
-        ...current,
-        ...patch,
-        updated_at: "2026-03-13T00:25:00Z",
-      };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  const recoveryEvents = await reconcileTrackedMergedButOpenIssues(
-    {
-      getPullRequestIfExists: async () => mergedPr,
-      getIssue: async () => {
-        getIssueCalls += 1;
-        return closedIssue;
-      },
-      closeIssue: async () => {
-        throw new Error("unexpected closeIssue call");
-      },
-      closePullRequest: async () => {
-        throw new Error("unexpected closePullRequest call");
-      },
-      getChecks: async () => [],
-      getMergedPullRequestsClosingIssue: async () => [],
-      getUnresolvedReviewThreads: async () => [],
-    },
-    stateStore,
-    state,
-    [],
-  );
-
-  assert.equal(getIssueCalls, 1);
-  assert.equal(saveCalls, 1);
-  assert.equal(state.issues["366"]?.state, "done");
-  assert.equal(state.issues["366"]?.pr_number, 191);
-  assert.equal(state.issues["366"]?.last_head_sha, "merged-head-191");
-  assert.deepEqual(recoveryEvents.map((event) => event.reason), [
-    "merged_pr_convergence: tracked PR #191 merged; marked issue #366 done",
-  ]);
-});
-
-test("reconcileTrackedMergedButOpenIssues does not rewrite recovery metadata when the done state is already current", async () => {
-  const original = createRecord({
-    issue_number: 366,
-    state: "done",
-    pr_number: 191,
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_blocker_signature: null,
-    last_failure_signature: null,
-    timeout_retry_count: 0,
-    blocked_verification_retry_count: 0,
-    repeated_blocker_count: 0,
-    repeated_failure_signature_count: 0,
-    last_head_sha: "merged-head-191",
-    local_review_blocker_summary: null,
-    local_review_recommendation: null,
-    local_review_degraded: false,
-    external_review_head_sha: null,
-    external_review_misses_path: null,
-    external_review_matched_findings_count: 0,
-    external_review_near_match_findings_count: 0,
-    external_review_missed_findings_count: 0,
-    last_recovery_reason: "existing recovery reason",
-    last_recovery_at: "2026-03-13T00:30:00Z",
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "366": original,
-    },
-  };
-  const mergedPr: GitHubPullRequest = {
-    number: 191,
-    title: "Merged implementation",
-    url: "https://example.test/pr/191",
-    state: "MERGED",
-    createdAt: "2026-03-13T00:10:00Z",
-    updatedAt: "2026-03-13T00:20:00Z",
-    isDraft: false,
-    headRefName: "codex/reopen-issue-366",
-    headRefOid: "merged-head-191",
-    mergeStateStatus: "CLEAN",
-    reviewDecision: "APPROVED",
-    mergedAt: "2026-03-13T00:20:00Z",
-    copilotReviewState: null,
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-  const closedIssue: GitHubIssue = {
-    number: 366,
-    title: "Merged implementation issue",
-    body: "",
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:21:00Z",
-    url: "https://example.test/issues/366",
-    state: "CLOSED",
-  };
-
-  let touchCalls = 0;
-  let saveCalls = 0;
-  const stateStore = {
-    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      touchCalls += 1;
-      return {
-        ...current,
-        ...patch,
-        updated_at: "2026-03-13T00:35:00Z",
-      };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  const recoveryEvents = await reconcileTrackedMergedButOpenIssues(
-    {
-      getPullRequestIfExists: async () => mergedPr,
-      getIssue: async () => closedIssue,
-      closeIssue: async () => {
-        throw new Error("unexpected closeIssue call");
-      },
-      closePullRequest: async () => {
-        throw new Error("unexpected closePullRequest call");
-      },
-      getChecks: async () => [],
-      getMergedPullRequestsClosingIssue: async () => [],
-      getUnresolvedReviewThreads: async () => [],
-    },
-    stateStore,
-    state,
-    [closedIssue],
-  );
-
-  assert.equal(touchCalls, 0);
-  assert.equal(saveCalls, 0);
-  assert.deepEqual(recoveryEvents, []);
-  assert.deepEqual(state.issues["366"], original);
-});
-
-test("runOnce recovers when post-codex refresh throws after leaving a dirty worktree", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 87;
-  const branch = branchName(fixture.config, issueNumber);
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "stabilizing",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        attempt_count: 1,
-        implementation_attempt_count: 1,
-        repair_attempt_count: 0,
-        last_error: null,
-        last_failure_kind: null,
-        last_failure_context: null,
-        last_failure_signature: null,
-        repeated_failure_signature_count: 0,
-        blocked_reason: null,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Reproduce dirty worktree recovery",
-    body: executionReadyBody("Reproduce dirty worktree recovery."),
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-
-  let resolveCalls = 0;
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async () => {
-      resolveCalls += 1;
-      if (resolveCalls === 1) {
-        return null;
-      }
-      throw new Error("post-turn refresh blew up");
-    },
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-  };
-
-  const message = await supervisor.runOnce({ dryRun: false });
-  assert.match(message, /Recovered from unexpected Codex turn failure/);
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(persisted.activeIssueNumber, null);
-  assert.equal(record.state, "failed");
-  assert.equal(record.last_failure_kind, "command_error");
-  assert.match(record.last_error ?? "", /post-turn refresh blew up/);
-  assert.equal(record.codex_session_id, "thread-123");
-  assert.match(record.last_codex_summary ?? "", /created a dirty checkpoint/);
-  assert.equal(record.blocked_reason, null);
-  assert.match(record.last_failure_context?.summary ?? "", /Supervisor failed while recovering a Codex turn/);
-  assert.deepEqual(record.last_failure_context?.details.slice(0, 4), [
-    "previous_state=stabilizing",
-    "workspace_dirty=yes",
-    `workspace_head=${record.last_head_sha}`,
-    "pr_number=none",
-  ]);
-
-  const worktreeStatus = git(["-C", path.join(fixture.workspaceRoot, `issue-${issueNumber}`), "status", "--short"]);
-  assert.match(worktreeStatus, /dirty\.txt/);
-
-  const issueLockPath = path.join(path.dirname(fixture.stateFile), "locks", "issues", `issue-${issueNumber}.lock`);
-  await assert.rejects(fs.access(issueLockPath));
-});
 
 test("runOnce records timeout bookkeeping when Codex exits non-zero", async () => {
   const fixture = await createSupervisorFixture({
@@ -1102,129 +123,6 @@ test("runOnce records timeout bookkeeping when Codex exits non-zero", async () =
   assert.match(record.last_error ?? "", /Command timed out after 1800000ms: codex exec/);
   assert.match(record.last_failure_context?.summary ?? "", /Codex exited non-zero for issue #89/);
   assert.match(record.last_failure_context?.details[0] ?? "", /Command timed out after 1800000ms: codex exec/);
-});
-
-test("recoverUnexpectedCodexTurnFailure preserves dirty recovery context and timeout bookkeeping", async () => {
-  const issueNumber = 88;
-  const record = createRecord({
-    issue_number: issueNumber,
-    state: "stabilizing",
-    timeout_retry_count: 1,
-    codex_session_id: "thread-456",
-    last_head_sha: "abc1234",
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: record,
-    },
-  };
-  let saveCalls = 0;
-  let syncedRecord: IssueRunRecord | null = null;
-  const stateStore = {
-    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      return { ...current, ...patch, updated_at: "2026-03-13T00:00:00.000Z" };
-    },
-    async save(): Promise<void> {
-      saveCalls += 1;
-    },
-  };
-
-  const updated = await recoverUnexpectedCodexTurnFailure({
-    stateStore: stateStore as unknown as Parameters<typeof recoverUnexpectedCodexTurnFailure>[0]["stateStore"],
-    state,
-    record,
-    issue: {
-      number: issueNumber,
-      title: "Timeout while recovering dirty worktree",
-      body: "",
-      createdAt: "2026-03-13T00:00:00Z",
-      updatedAt: "2026-03-13T00:00:00Z",
-      url: `https://example.test/issues/${issueNumber}`,
-      state: "OPEN",
-    },
-    journalSync: async (nextRecord) => {
-      syncedRecord = nextRecord;
-    },
-    error: new Error("Command timed out after 1800000ms: codex exec resume thread-456"),
-    workspaceStatus: {
-      hasUncommittedChanges: true,
-      headSha: "deadbee",
-    },
-    pr: {
-      number: 55,
-      headRefOid: "feed123",
-    },
-  });
-
-  assert.equal(saveCalls, 1);
-  assert.equal(state.activeIssueNumber, null);
-  assert.equal(updated.state, "failed");
-  assert.equal(updated.last_failure_kind, "timeout");
-  assert.equal(updated.timeout_retry_count, 2);
-  assert.equal(updated.blocked_reason, null);
-  assert.match(updated.last_error ?? "", /Command timed out after 1800000ms/);
-  assert.match(updated.last_failure_context?.summary ?? "", /Supervisor failed while recovering a Codex turn/);
-  assert.deepEqual(updated.last_failure_context?.details.slice(0, 6), [
-    "previous_state=stabilizing",
-    "workspace_dirty=yes",
-    "workspace_head=deadbee",
-    "pr_number=55",
-    "pr_head=feed123",
-    "codex_session_id=thread-456",
-  ]);
-  assert.equal(state.issues[String(issueNumber)], updated);
-  assert.equal(syncedRecord, updated);
-});
-
-test("recoverUnexpectedCodexTurnFailure records unavailable workspace inspection distinctly", async () => {
-  const issueNumber = 92;
-  const record = createRecord({
-    issue_number: issueNumber,
-    state: "stabilizing",
-    last_head_sha: "abc1234",
-    codex_session_id: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: record,
-    },
-  };
-  const stateStore = {
-    touch(current: IssueRunRecord, patch: Partial<IssueRunRecord>): IssueRunRecord {
-      return { ...current, ...patch, updated_at: "2026-03-15T02:00:00.000Z" };
-    },
-    async save(): Promise<void> {},
-  };
-
-  const updated = await recoverUnexpectedCodexTurnFailure({
-    stateStore: stateStore as unknown as Parameters<typeof recoverUnexpectedCodexTurnFailure>[0]["stateStore"],
-    state,
-    record,
-    issue: {
-      number: issueNumber,
-      title: "Refresh failed before workspace inspection",
-      body: "",
-      createdAt: "2026-03-15T00:00:00Z",
-      updatedAt: "2026-03-15T00:00:00Z",
-      url: `https://example.test/issues/${issueNumber}`,
-      state: "OPEN",
-    },
-    journalSync: async () => {},
-    error: new Error("Command failed: codex exec resume"),
-    workspaceStatus: null,
-    pr: null,
-  });
-
-  assert.deepEqual(updated.last_failure_context?.details.slice(0, 6), [
-    "previous_state=stabilizing",
-    "workspace_dirty=unknown",
-    "workspace_head=abc1234",
-    "pr_number=none",
-    "pr_head=none",
-    "codex_session_id=none",
-  ]);
 });
 
 test("handleAuthFailure blocks the active issue and preserves failure tracking fields", async () => {
@@ -2241,1080 +1139,6 @@ Decide whether to keep the current production auth token flow or replace it befo
   const journal = await fs.readFile(record.journal_path ?? "", "utf8");
   assert.match(journal, /requires manual clarification because high-risk blocking ambiguity/i);
   assert.match(journal, /ambiguity_classes=unresolved_choice/i);
-});
-
-test("status shows readiness reasons for runnable, requirements-blocked, and clarification-blocked issues", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "91": createRecord({
-        issue_number: 91,
-        state: "done",
-        branch: branchName(fixture.config, 91),
-        workspace: path.join(fixture.workspaceRoot, "issue-91"),
-        journal_path: null,
-        blocked_reason: null,
-        last_error: null,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const runnableIssue: GitHubIssue = {
-    number: 92,
-    title: "Step 2",
-    body: `## Summary
-Ship the second step.
-
-## Scope
-- build on the completed dependency
-
-## Acceptance criteria
-- supervisor can explain why this issue is runnable
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Depends on: #91`,
-    createdAt: "2026-03-13T00:05:00Z",
-    updatedAt: "2026-03-13T00:05:00Z",
-    url: "https://example.test/issues/92",
-    state: "OPEN",
-  };
-  const missingMetadataIssue: GitHubIssue = {
-    number: 93,
-    title: "Underspecified issue",
-    body: `## Summary
-Missing execution-ready metadata.`,
-    createdAt: "2026-03-13T00:10:00Z",
-    updatedAt: "2026-03-13T00:10:00Z",
-    url: "https://example.test/issues/93",
-    state: "OPEN",
-  };
-  const clarificationBlockedIssue: GitHubIssue = {
-    number: 94,
-    title: "Decide which auth path to keep",
-    body: `## Summary
-Decide whether to keep the current production auth token flow or replace it before rollout.
-
-## Scope
-- choose the production authentication path for service-to-service traffic
-
-## Acceptance criteria
-- the operator confirms which auth flow should ship
-
-## Verification
-- npm test -- src/supervisor.test.ts`,
-    createdAt: "2026-03-13T00:15:00Z",
-    updatedAt: "2026-03-13T00:15:00Z",
-    url: "https://example.test/issues/94",
-    state: "OPEN",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    listCandidateIssues: async () => [runnableIssue, missingMetadataIssue, clarificationBlockedIssue],
-    getPullRequestIfExists: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(status, /runnable_issues=#92 ready=execution_ready\+depends_on_satisfied:91/);
-  assert.match(
-    status,
-    /blocked_issues=#93 blocked_by=requirements:scope, acceptance criteria, verification; #94 blocked_by=clarification:unresolved_choice:auth/,
-  );
-});
-
-test("status marks skipped readiness checks explicitly and uses non-conflicting inner separators", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "91": createRecord({
-        issue_number: 91,
-        state: "done",
-        branch: branchName(fixture.config, 91),
-        workspace: path.join(fixture.workspaceRoot, "issue-91"),
-        journal_path: null,
-        blocked_reason: null,
-        last_error: null,
-      }),
-      "92": createRecord({
-        issue_number: 92,
-        state: "done",
-        branch: branchName(fixture.config, 92),
-        workspace: path.join(fixture.workspaceRoot, "issue-92"),
-        journal_path: null,
-        blocked_reason: null,
-        last_error: null,
-      }),
-      "93": createRecord({
-        issue_number: 93,
-        state: "queued",
-        branch: branchName(fixture.config, 93),
-        workspace: path.join(fixture.workspaceRoot, "issue-93"),
-        journal_path: null,
-        blocked_reason: null,
-        last_error: null,
-        attempt_count: 1,
-        implementation_attempt_count: 1,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const predecessorOne: GitHubIssue = {
-    number: 91,
-    title: "Step 1",
-    body: `## Summary
-Finish step 1.
-
-## Scope
-- start the execution order chain
-
-## Acceptance criteria
-- step 1 completes first
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Part of: #150
-Execution order: 1 of 3`,
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: "https://example.test/issues/91",
-    state: "CLOSED",
-  };
-  const predecessorTwo: GitHubIssue = {
-    number: 92,
-    title: "Step 2",
-    body: `## Summary
-Finish step 2.
-
-## Scope
-- land after step 1
-
-## Acceptance criteria
-- step 2 completes after step 1
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Part of: #150
-Execution order: 2 of 3`,
-    createdAt: "2026-03-13T00:05:00Z",
-    updatedAt: "2026-03-13T00:05:00Z",
-    url: "https://example.test/issues/92",
-    state: "CLOSED",
-  };
-  const skippedRequirementsIssue: GitHubIssue = {
-    number: 93,
-    title: "Step 3",
-    body: `## Summary
-Existing in-flight issue with missing readiness metadata.
-
-Depends on: #91, #92
-Part of: #150
-Execution order: 3 of 3`,
-    createdAt: "2026-03-13T00:10:00Z",
-    updatedAt: "2026-03-13T00:10:00Z",
-    url: "https://example.test/issues/93",
-    state: "OPEN",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    listCandidateIssues: async () => [predecessorOne, predecessorTwo, skippedRequirementsIssue],
-    getPullRequestIfExists: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(
-    status,
-    /runnable_issues=#93 ready=requirements_skipped\+depends_on_satisfied:91\|92\+execution_order_satisfied:91\|92/,
-  );
-});
-
-test("status --why explains why the current runnable issue was selected", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "91": createRecord({
-        issue_number: 91,
-        state: "done",
-        branch: branchName(fixture.config, 91),
-        workspace: path.join(fixture.workspaceRoot, "issue-91"),
-        journal_path: null,
-        blocked_reason: null,
-        last_error: null,
-      }),
-      "92": createRecord({
-        issue_number: 92,
-        state: "done",
-        branch: branchName(fixture.config, 92),
-        workspace: path.join(fixture.workspaceRoot, "issue-92"),
-        journal_path: null,
-        blocked_reason: null,
-        last_error: null,
-      }),
-      "95": createRecord({
-        issue_number: 95,
-        state: "blocked",
-        branch: branchName(fixture.config, 95),
-        workspace: path.join(fixture.workspaceRoot, "issue-95"),
-        journal_path: null,
-        blocked_reason: "verification",
-        last_error: "verification still failing",
-        blocked_verification_retry_count: 1,
-        repeated_blocker_count: fixture.config.sameBlockerRepeatLimit,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const blockedIssue: GitHubIssue = {
-    number: 95,
-    title: "Blocked verification retry",
-    body: `## Summary
-Retry the failing verification.
-
-## Scope
-- rerun the failing check
-
-## Acceptance criteria
-- verification can pass
-
-## Verification
-- npm test -- src/supervisor.test.ts`,
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: "https://example.test/issues/95",
-    state: "OPEN",
-  };
-  const predecessorIssueOne: GitHubIssue = {
-    number: 91,
-    title: "Step 1",
-    body: `## Summary
-Ship the first step.
-
-## Scope
-- start the execution order chain
-
-## Acceptance criteria
-- step one lands first
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Part of: #150
-Execution order: 1 of 3`,
-    createdAt: "2026-03-12T23:55:00Z",
-    updatedAt: "2026-03-12T23:55:00Z",
-    url: "https://example.test/issues/91",
-    state: "CLOSED",
-  };
-  const predecessorIssueTwo: GitHubIssue = {
-    number: 92,
-    title: "Step 2",
-    body: `## Summary
-Ship the second step.
-
-## Scope
-- continue the execution order chain
-
-## Acceptance criteria
-- step two lands after step one
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Part of: #150
-Execution order: 2 of 3`,
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: "https://example.test/issues/92",
-    state: "CLOSED",
-  };
-  const selectedIssue: GitHubIssue = {
-    number: 93,
-    title: "Step 3",
-    body: `## Summary
-Ship the third step.
-
-## Scope
-- build after the first two steps land
-
-## Acceptance criteria
-- status explains why this issue is selected
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Depends on: #91
-Part of: #150
-Execution order: 3 of 3`,
-    createdAt: "2026-03-13T00:05:00Z",
-    updatedAt: "2026-03-13T00:05:00Z",
-    url: "https://example.test/issues/93",
-    state: "OPEN",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    listAllIssues: async () => [predecessorIssueOne, predecessorIssueTwo, blockedIssue, selectedIssue],
-    listCandidateIssues: async () => [blockedIssue, selectedIssue],
-    getPullRequestIfExists: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status({ why: true });
-
-  assert.match(status, /selected_issue=#93/);
-  assert.match(
-    status,
-    /selection_reason=ready execution_ready=yes depends_on=91:done execution_order=150\/3 predecessors=91\|92:done retry_state=fresh/,
-  );
-});
-
-test("explain reports dependency blockers for a non-runnable issue", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {},
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const dependencyIssue: GitHubIssue = {
-    number: 91,
-    title: "Step 1",
-    body: `## Summary
-Ship the first step.
-
-## Scope
-- land the dependency first
-
-## Acceptance criteria
-- step one completes before step two
-
-## Verification
-- npm test -- src/supervisor.test.ts`,
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: "https://example.test/issues/91",
-    state: "OPEN",
-  };
-  const blockedIssue: GitHubIssue = {
-    number: 93,
-    title: "Step 2",
-    body: `## Summary
-Ship the second step.
-
-## Scope
-- wait for the dependency to finish first
-
-## Acceptance criteria
-- explain shows the dependency gate
-
-## Verification
-- npm test -- src/supervisor.test.ts
-
-Depends on: #91`,
-    createdAt: "2026-03-13T00:05:00Z",
-    updatedAt: "2026-03-13T00:05:00Z",
-    url: "https://example.test/issues/93",
-    state: "OPEN",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    getIssue: async () => blockedIssue,
-    listAllIssues: async () => [dependencyIssue, blockedIssue],
-    listCandidateIssues: async () => [dependencyIssue, blockedIssue],
-  };
-
-  const explanation = await supervisor.explain(93);
-
-  assert.match(explanation, /^issue=#93$/m);
-  assert.match(explanation, /^state=untracked$/m);
-  assert.match(explanation, /^runnable=no$/m);
-  assert.match(explanation, /^reason_1=dependency depends on #91$/m);
-});
-
-test("explain reports candidate filtering for a non-candidate issue", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {},
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const filteredIssue: GitHubIssue = {
-    number: 94,
-    title: "Filtered out of candidate selection",
-    body: executionReadyBody("Explain should report when scheduler filters out the issue."),
-    createdAt: "2026-03-13T00:05:00Z",
-    updatedAt: "2026-03-13T00:05:00Z",
-    url: "https://example.test/issues/94",
-    state: "CLOSED",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    getIssue: async () => filteredIssue,
-    listAllIssues: async () => [filteredIssue],
-    listCandidateIssues: async () => [],
-  };
-
-  const explanation = await supervisor.explain(94);
-
-  assert.match(explanation, /^issue=#94$/m);
-  assert.match(explanation, /^state=untracked$/m);
-  assert.match(explanation, /^runnable=no$/m);
-  assert.match(explanation, /^reason_1=candidate filtered_by_candidate_list$/m);
-});
-
-test("explain reports retry-budget blockers for verification-blocked issues", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "95": createRecord({
-        issue_number: 95,
-        state: "blocked",
-        branch: branchName(fixture.config, 95),
-        workspace: path.join(fixture.workspaceRoot, "issue-95"),
-        journal_path: null,
-        blocked_reason: "verification",
-        last_error: "verification still failing",
-        blocked_verification_retry_count: fixture.config.blockedVerificationRetryLimit,
-        repeated_blocker_count: 1,
-        repeated_failure_signature_count: 1,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const blockedIssue: GitHubIssue = {
-    number: 95,
-    title: "Blocked verification retry",
-    body: `## Summary
-Retry the failing verification.
-
-## Scope
-- rerun the failing check
-
-## Acceptance criteria
-- verification can pass
-
-## Verification
-- npm test -- src/supervisor.test.ts`,
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: "https://example.test/issues/95",
-    state: "OPEN",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    getIssue: async () => blockedIssue,
-    listAllIssues: async () => [blockedIssue],
-    listCandidateIssues: async () => [blockedIssue],
-  };
-
-  const explanation = await supervisor.explain(95);
-
-  assert.match(explanation, /^state=blocked$/m);
-  assert.match(explanation, /^blocked_reason=verification$/m);
-  assert.match(explanation, /^runnable=no$/m);
-  assert.match(
-    explanation,
-    new RegExp(`^reason_1=retry_budget blocked_verification_retry_count=${fixture.config.blockedVerificationRetryLimit}\\/${fixture.config.blockedVerificationRetryLimit}$`, "m"),
-  );
-  assert.match(explanation, /^reason_2=local_state blocked$/m);
-  assert.match(explanation, /^last_error=verification still failing$/m);
-});
-
-test("explain reports manual review blockers for blocked issues", async () => {
-  const fixture = await createSupervisorFixture();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "97": createRecord({
-        issue_number: 97,
-        state: "blocked",
-        branch: branchName(fixture.config, 97),
-        workspace: path.join(fixture.workspaceRoot, "issue-97"),
-        journal_path: null,
-        blocked_reason: "manual_review",
-        last_error: "waiting on human review",
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const blockedIssue: GitHubIssue = {
-    number: 97,
-    title: "Manual review blocker",
-    body: `## Summary
-Wait for a human review before proceeding.
-
-## Scope
-- hold the rollout until the reviewer signs off
-
-## Acceptance criteria
-- explain shows the manual block reason
-
-## Verification
-- npm test -- src/supervisor.test.ts`,
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: "https://example.test/issues/97",
-    state: "OPEN",
-  };
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    getIssue: async () => blockedIssue,
-    listAllIssues: async () => [blockedIssue],
-    listCandidateIssues: async () => [blockedIssue],
-  };
-
-  const explanation = await supervisor.explain(97);
-
-  assert.match(explanation, /^state=blocked$/m);
-  assert.match(explanation, /^blocked_reason=manual_review$/m);
-  assert.match(explanation, /^runnable=no$/m);
-  assert.match(explanation, /^reason_1=manual_block manual_review$/m);
-  assert.match(explanation, /^reason_2=local_state blocked$/m);
-});
-
-test("status includes a compact handoff summary for an active blocker", async () => {
-  const fixture = await createSupervisorFixture();
-  const journalPath = path.join(fixture.workspaceRoot, "issue-92", ".codex-supervisor", "issue-journal.md");
-  await fs.mkdir(path.dirname(journalPath), { recursive: true });
-  await fs.writeFile(
-    journalPath,
-    `# Issue #92: Step 2
-
-## Supervisor Snapshot
-- Updated at: 2026-03-13T00:20:00Z
-
-## Latest Codex Summary
-- None yet.
-
-## Active Failure Context
-- None recorded.
-
-## Codex Working Notes
-### Current Handoff
-- Hypothesis: The status output should summarize the live handoff.
-- What changed: Added structured journal fields.
-- Current blocker: Waiting on the status formatter to show the blocker and next step.
-- Next exact step: Render a compact handoff summary in status output.
-- Verification gap: Focused supervisor status test still missing.
-- Files touched: src/journal.ts, src/supervisor.ts
-- Rollback concern:
-- Last focused command: npm test -- --test-name-pattern handoff
-
-### Scratchpad
-- Keep this section short.
-`,
-    "utf8",
-  );
-
-  const activeRecord = createRecord({
-    issue_number: 92,
-    state: "reproducing",
-    branch: branchName(fixture.config, 92),
-    workspace: path.join(fixture.workspaceRoot, "issue-92"),
-    journal_path: journalPath,
-    blocked_reason: "verification",
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 92,
-    issues: {
-      "92": activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(
-    status,
-    /handoff_summary=blocker: Waiting on the status formatter to show the blocker and next step\. \| next: Render a compact handoff summary in status output\./,
-  );
-});
-
-test("status keeps the active handoff summary when PR status loading emits a warning", async () => {
-  const fixture = await createSupervisorFixture();
-  const journalPath = path.join(fixture.workspaceRoot, "issue-92", ".codex-supervisor", "issue-journal.md");
-  await fs.mkdir(path.dirname(journalPath), { recursive: true });
-  await fs.writeFile(
-    journalPath,
-    `# Issue #92: Step 2
-
-## Supervisor Snapshot
-- Updated at: 2026-03-13T00:20:00Z
-
-## Latest Codex Summary
-- None yet.
-
-## Active Failure Context
-- None recorded.
-
-## Codex Working Notes
-### Current Handoff
-- Hypothesis: Preserve the active handoff summary even when status loading warns.
-- What changed: Added a focused status warning assertion.
-- Current blocker: Waiting on GitHub status hydration to finish cleanly.
-- Next exact step: Keep the warning path rendering the same handoff summary.
-- Verification gap: Focused supervisor status warning coverage was missing.
-- Files touched: src/supervisor.test.ts
-- Rollback concern:
-- Last focused command: npm test -- --test-name-pattern status warning
-
-### Scratchpad
-- Keep this section short.
-`,
-    "utf8",
-  );
-
-  const activeRecord = createRecord({
-    issue_number: 92,
-    state: "reproducing",
-    branch: branchName(fixture.config, 92),
-    workspace: path.join(fixture.workspaceRoot, "issue-92"),
-    journal_path: journalPath,
-    blocked_reason: "verification",
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 92,
-    issues: {
-      "92": activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => {
-      throw new Error("injected status hydration failure");
-    },
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(
-    status,
-    /handoff_summary=blocker: Waiting on GitHub status hydration to finish cleanly\. \| next: Keep the warning path rendering the same handoff summary\./,
-  );
-  assert.match(status, /status_warning=injected status hydration failure/);
-});
-
-test("status downgrades journal read failures into status warnings", async () => {
-  const fixture = await createSupervisorFixture();
-  const journalPath = path.join(fixture.workspaceRoot, "issue-92");
-  await fs.mkdir(journalPath, { recursive: true });
-
-  const activeRecord = createRecord({
-    issue_number: 92,
-    state: "reproducing",
-    branch: branchName(fixture.config, 92),
-    workspace: path.join(fixture.workspaceRoot, "issue-92-workspace"),
-    journal_path: journalPath,
-    blocked_reason: "verification",
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 92,
-    issues: {
-      "92": activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(status, /status_warning=/);
-  assert.doesNotMatch(status, /handoff_summary=/);
-});
-
-test("status shows durable guardrail provenance for active committed and runtime guidance", async () => {
-  const fixture = await createSupervisorFixture();
-  fixture.config.localReviewArtifactDir = path.join(path.dirname(fixture.stateFile), "reviews");
-
-  const issueNumber = 92;
-  const branch = branchName(fixture.config, issueNumber);
-  git(["checkout", "-b", branch], fixture.repoPath);
-  await fs.mkdir(path.join(fixture.repoPath, "src"), { recursive: true });
-  await fs.writeFile(
-    path.join(fixture.repoPath, "src", "auth.ts"),
-    "export function canUpdateRecord(): boolean {\n  return true;\n}\n",
-    "utf8",
-  );
-  git(["add", "src/auth.ts"], fixture.repoPath);
-  git(["commit", "-m", "Add auth change"], fixture.repoPath);
-  const headSha = git(["rev-parse", "HEAD"], fixture.repoPath);
-
-  await fs.mkdir(path.join(fixture.repoPath, "docs", "shared-memory"), { recursive: true });
-  await fs.writeFile(
-    path.join(fixture.repoPath, "docs", "shared-memory", "verifier-guardrails.json"),
-    `${JSON.stringify({
-      version: 1,
-      rules: [
-        {
-          id: "auth-direct-guard",
-          title: "Re-check auth guard changes directly",
-          file: "src/auth.ts",
-          line: 1,
-          summary: "Auth guard changes must be re-read directly before dismissing high-severity findings.",
-          rationale: "A prior verifier miss cleared an auth fallback without inspecting the guard path itself.",
-        },
-      ],
-    }, null, 2)}\n`,
-    "utf8",
-  );
-  await fs.writeFile(
-    path.join(fixture.repoPath, "docs", "shared-memory", "external-review-guardrails.json"),
-    `${JSON.stringify({
-      version: 1,
-      patterns: [
-        {
-          fingerprint: "committed-auth-guard",
-          reviewerLogin: "copilot-pull-request-reviewer",
-          file: "src/auth.ts",
-          line: 1,
-          summary: "Permission checks in auth flows deserve an explicit local-review pass.",
-          rationale: "A committed external review miss showed auth guard regressions were previously skipped.",
-          sourceArtifactPath: "owner-repo/issue-12/external-review-misses-head-aaaabbbbcccc.json",
-          sourceHeadSha: "aaaabbbbccccdddd",
-          lastSeenAt: "2026-03-10T00:00:00Z",
-        },
-      ],
-    }, null, 2)}\n`,
-    "utf8",
-  );
-
-  const artifactDir = path.join(fixture.config.localReviewArtifactDir, "owner-repo", `issue-${issueNumber}`);
-  await fs.mkdir(artifactDir, { recursive: true });
-  await fs.writeFile(
-    path.join(artifactDir, "external-review-misses-head-111122223333.json"),
-    `${JSON.stringify({
-      issueNumber,
-      prNumber: 44,
-      branch,
-      headSha: "1111222233334444",
-      generatedAt: "2026-03-12T00:00:00Z",
-      findings: [],
-      reusableMissPatterns: [
-        {
-          fingerprint: "runtime-auth-guard",
-          reviewerLogin: "copilot-pull-request-reviewer",
-          file: "src/auth.ts",
-          line: 2,
-          summary: "Runtime artifact keeps auth fallback blind spots active until local review covers them.",
-          rationale: "A recent external review still found the fallback path unreviewed locally.",
-          sourceArtifactPath: path.join(artifactDir, "external-review-misses-head-111122223333.json"),
-          sourceHeadSha: "1111222233334444",
-          lastSeenAt: "2026-03-12T00:00:00Z",
-        },
-      ],
-      durableGuardrailCandidates: [],
-      regressionTestCandidates: [],
-      counts: {
-        matched: 0,
-        nearMatch: 0,
-        missedByLocalReview: 1,
-      },
-    }, null, 2)}\n`,
-    "utf8",
-  );
-
-  const activeRecord = createRecord({
-    issue_number: issueNumber,
-    state: "reproducing",
-    branch,
-    workspace: fixture.repoPath,
-    journal_path: null,
-    blocked_reason: null,
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => ({
-      number: 44,
-      title: "Auth guard",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-13T00:00:00Z",
-      isDraft: true,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: branch,
-      headRefOid: headSha,
-      mergedAt: null,
-    }),
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(
-    status,
-    /durable_guardrails verifier=committed:docs\/shared-memory\/verifier-guardrails\.json#1 external_review=committed:docs\/shared-memory\/external-review-guardrails\.json#1\|runtime:owner-repo\/issue-92\/external-review-misses-head-111122223333\.json#1/,
-  );
-});
-
-test("status guardrail provenance reflects the merged active external-review winners", async () => {
-  const fixture = await createSupervisorFixture();
-  fixture.config.localReviewArtifactDir = path.join(path.dirname(fixture.stateFile), "reviews");
-
-  const issueNumber = 92;
-  const branch = branchName(fixture.config, issueNumber);
-  git(["checkout", "-b", branch], fixture.repoPath);
-  await fs.mkdir(path.join(fixture.repoPath, "src"), { recursive: true });
-  await fs.writeFile(
-    path.join(fixture.repoPath, "src", "auth.ts"),
-    "export function canUpdateRecord(): boolean {\n  return true;\n}\n",
-    "utf8",
-  );
-  git(["add", "src/auth.ts"], fixture.repoPath);
-  git(["commit", "-m", "Add auth change"], fixture.repoPath);
-  const headSha = git(["rev-parse", "HEAD"], fixture.repoPath);
-
-  await fs.mkdir(path.join(fixture.repoPath, "docs", "shared-memory"), { recursive: true });
-  await fs.writeFile(
-    path.join(fixture.repoPath, "docs", "shared-memory", "external-review-guardrails.json"),
-    `${JSON.stringify({
-      version: 1,
-      patterns: [
-        {
-          fingerprint: "shared-auth-guard",
-          reviewerLogin: "copilot-pull-request-reviewer",
-          file: "src/auth.ts",
-          line: 1,
-          summary: "Committed auth guard guidance.",
-          rationale: "Older committed guidance for the same auth blind spot.",
-          sourceArtifactPath: "owner-repo/issue-12/external-review-misses-head-aaaabbbbcccc.json",
-          sourceHeadSha: "aaaabbbbccccdddd",
-          lastSeenAt: "2026-03-10T00:00:00Z",
-        },
-      ],
-    }, null, 2)}\n`,
-    "utf8",
-  );
-
-  const artifactDir = path.join(fixture.config.localReviewArtifactDir, "owner-repo", `issue-${issueNumber}`);
-  await fs.mkdir(artifactDir, { recursive: true });
-  await fs.writeFile(
-    path.join(artifactDir, "external-review-misses-head-111122223333.json"),
-    `${JSON.stringify({
-      issueNumber,
-      prNumber: 44,
-      branch,
-      headSha: "1111222233334444",
-      generatedAt: "2026-03-12T00:00:00Z",
-      findings: [],
-      reusableMissPatterns: [
-        {
-          fingerprint: "shared-auth-guard",
-          reviewerLogin: "copilot-pull-request-reviewer",
-          file: "src/auth.ts",
-          line: 2,
-          summary: "Runtime auth guard guidance.",
-          rationale: "Newer runtime guidance for the same auth blind spot should win.",
-          sourceArtifactPath: path.join(artifactDir, "external-review-misses-head-111122223333.json"),
-          sourceHeadSha: "1111222233334444",
-          lastSeenAt: "2026-03-12T00:00:00Z",
-        },
-      ],
-      durableGuardrailCandidates: [],
-      regressionTestCandidates: [],
-      counts: {
-        matched: 0,
-        nearMatch: 0,
-        missedByLocalReview: 1,
-      },
-    }, null, 2)}\n`,
-    "utf8",
-  );
-
-  const activeRecord = createRecord({
-    issue_number: issueNumber,
-    state: "addressing_review",
-    branch,
-    workspace: fixture.repoPath,
-    journal_path: null,
-    blocked_reason: null,
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => ({
-      number: 44,
-      title: "Auth guard",
-      url: "https://example.test/pr/44",
-      state: "OPEN",
-      createdAt: "2026-03-13T00:00:00Z",
-      isDraft: false,
-      reviewDecision: null,
-      mergeStateStatus: "CLEAN",
-      mergeable: "MERGEABLE",
-      headRefName: branch,
-      headRefOid: headSha,
-      mergedAt: null,
-    }),
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.match(
-    status,
-    /durable_guardrails verifier=none external_review=runtime:owner-repo\/issue-92\/external-review-misses-head-111122223333\.json#1/,
-  );
-  assert.doesNotMatch(status, /external_review=committed:/);
-});
-
-test("status omits durable guardrail warnings when the workspace diff cannot be read", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 92;
-  const activeRecord = createRecord({
-    issue_number: issueNumber,
-    state: "addressing_review",
-    branch: branchName(fixture.config, issueNumber),
-    workspace: path.join(fixture.workspaceRoot, "missing-workspace"),
-    journal_path: null,
-    blocked_reason: null,
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.doesNotMatch(status, /durable_guardrails /);
-  assert.doesNotMatch(status, /status_warning=/);
-});
-
-test("status omits handoff summary when the handoff has no actionable blocker or next step", async () => {
-  const fixture = await createSupervisorFixture();
-  const journalPath = path.join(fixture.workspaceRoot, "issue-92", ".codex-supervisor", "issue-journal.md");
-  await fs.mkdir(path.dirname(journalPath), { recursive: true });
-  await fs.writeFile(
-    journalPath,
-    `# Issue #92: Step 2
-
-## Supervisor Snapshot
-- Updated at: 2026-03-13T00:20:00Z
-
-## Latest Codex Summary
-- None yet.
-
-## Active Failure Context
-- None recorded.
-
-## Codex Working Notes
-### Current Handoff
-- Hypothesis:
-- What changed: Added structured journal fields.
-- Current blocker: None.
-- Next exact step:
-- Verification gap: None.
-- Files touched: src/journal.ts
-- Rollback concern:
-- Last focused command: npm test -- --test-name-pattern handoff
-
-### Scratchpad
-- Keep this section short.
-`,
-    "utf8",
-  );
-
-  const activeRecord = createRecord({
-    issue_number: 92,
-    state: "implementing",
-    branch: branchName(fixture.config, 92),
-    workspace: path.join(fixture.workspaceRoot, "issue-92"),
-    journal_path: journalPath,
-    blocked_reason: null,
-    last_error: null,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: 92,
-    issues: {
-      "92": activeRecord,
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    resolvePullRequestForBranch: async () => null,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-  };
-
-  const status = await supervisor.status();
-
-  assert.doesNotMatch(status, /handoff_summary=/);
 });
 
 test("runOnce still prefers a ready issue over dependency-blocked candidates", async () => {
@@ -4841,10 +2665,6 @@ test("runOnce routes supervisor turn execution through an injected agent runner"
   assert.equal(record.last_failure_kind, null);
   assert.match(record.last_codex_summary ?? "", /completed via injected agent runner/);
 });
-
-function branchName(config: SupervisorConfig, issueNumber: number): string {
-  return `${config.branchPrefix}${issueNumber}`;
-}
 
 test("supervisor module continues to export the Supervisor class", () => {
   assert.equal(typeof Supervisor, "function");
